+++ Apps/WebClient/src/ClientApp/src/views/reports.vue
@@ -1,220 +1,58 @@
 <script lang="ts">
 import Vue from "vue";
 import { Component, Ref } from "vue-property-decorator";
-import { Action, Getter } from "vuex-class";
 import PageTitleComponent from "@/components/pageTitle.vue";
 import { ILogger } from "@/services/interfaces";
 import container from "@/plugins/inversify.config";
 import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
-import BannerError from "@/models/bannerError";
 import MessageModalComponent from "@/components/modal/genericMessage.vue";
 import MedicationHistoryReportComponent from "@/components/report/medicationHistory.vue";
-import LoadingComponent from "@/components/loading.vue";
-import ProtectiveWordComponent from "@/components/modal/protectiveWord.vue";
-import MedicationStatementHistory from "@/models/medicationStatementHistory";
-import RequestResult from "@/models/requestResult";
-import User from "@/models/user";
-import { ResultType } from "@/constants/resulttype";
-import ErrorTranslator from "@/utility/errorTranslator";
+import MSPVisitsReportComponent from "@/components/report/mspVisits.vue";
 import { IAuthenticationService } from "@/services/interfaces";
-import html2pdf from "html2pdf.js";
-
-/**
- * Shallow representation to be used by html2pdf. Does not define the object complety, but sufficiently to avoid using any.
- * To be expanded as needed.
- */
-interface PDFDefinition {
-    internal: {
-        getNumberOfPages(): number;
-        pageSize: { getWidth(): number; getHeight(): number };
-    };
-    setPage(page: number): void;
-    setFontSize(size: number): void;
-    setTextColor(color: number): void;
-    text(text: string, width: number, height: number): void;
-}
 
 @Component({
     components: {
         PageTitleComponent,
         MessageModalComponent,
         MedicationHistoryReportComponent,
-        LoadingComponent,
-        ProtectiveWordComponent,
+        MSPVisitsReportComponent,
     },
 })
 export default class ReportsView extends Vue {
-    @Ref("report")
-    readonly report!: HTMLElement;
     @Ref("messageModal")
     readonly messageModal!: MessageModalComponent;
-    @Ref("protectiveWordModal")
-    readonly protectiveWordModal!: ProtectiveWordComponent;
-    @Getter("user", { namespace: "user" })
-    private user!: User;
-    @Action("getMedicationStatements", { namespace: "medication" })
-    private getMedicationStatements!: (params: {
-        hdid: string;
-        protectiveWord?: string;
-    }) => Promise<RequestResult<MedicationStatementHistory[]>>;
-    @Action("addError", { namespace: "errorBanner" })
-    private addError!: (error: BannerError) => void;
+    @Ref("medicationHistoryReport")
+    readonly medicationHistoryReport!: MedicationHistoryReportComponent;
+    @Ref("mspVisitsReport")
+    readonly mspVisitsReport!: MSPVisitsReportComponent;
 
     private fullName = "";
-    private medicationStatementHistoryPage: MedicationStatementHistory[] = [];
-    private medicationStatementHistory: MedicationStatementHistory[] = [];
-    private isLoading = false;
-    private protectiveWordAttempts = 0;
     private logger!: ILogger;
-    private isDataLoaded = false;
-    private fileMaxRecords = 1000;
     private reportType = "";
     private reportTypeOptions = [
         { value: "", text: "Select a service" },
         { value: "MED", text: "Medications" },
+        { value: "MSP", text: "MSP Visits" },
     ];
-    private get totalFiles(): number {
-        return Math.ceil(
-            this.medicationStatementHistory.length / this.fileMaxRecords
-        );
-    }
 
     private showConfirmationModal() {
         this.messageModal.showModal();
     }
 
-    private async generateMedicationHistoryPdf(fileIndex = 0) {
-        this.logger.debug("generating Medication History PDF...");
-        this.isLoading = true;
-
-        // Breaks records into chunks for multiple files.
-        this.medicationStatementHistoryPage = this.medicationStatementHistory.slice(
-            fileIndex * this.fileMaxRecords,
-            (fileIndex + 1) * this.fileMaxRecords
-        );
-
-        let opt = {
-            margin: [15, 15],
-            filename: `HealthGateway_MedicationHistory_File${
-                fileIndex + 1
-            }.pdf`,
-            image: { type: "jpeg", quality: 1 },
-            html2canvas: { dpi: 192, scale: 1.2, letterRendering: true },
-            jsPDF: { unit: "pt", format: "letter", orientation: "portrait" },
-            pagebreak: { mode: ["avoid-all", "css", "legacy"] },
-        };
-        html2pdf()
-            .set(opt)
-            .from(this.report)
-            .toPdf()
-            .get("pdf")
-            .then((pdf: PDFDefinition) => {
-                // Add footer with page numbers
-                var totalPages = pdf.internal.getNumberOfPages();
-                for (let i = 1; i <= totalPages; i++) {
-                    pdf.setPage(i);
-                    pdf.setFontSize(10);
-                    pdf.setTextColor(150);
-                    pdf.text(
-                        `Page ${i} of ${totalPages} - File ${
-                            fileIndex + 1
-                        } of ${this.totalFiles}`,
-                        pdf.internal.pageSize.getWidth() / 2 - 55,
-                        pdf.internal.pageSize.getHeight() - 10
-                    );
-                }
-            })
-            .save()
-            .output("bloburl")
-            .then((pdfBlobUrl: RequestInfo) => {
-                fetch(pdfBlobUrl).then((res) => {
-                    res.blob().then(() => {
-                        if (fileIndex + 1 < this.totalFiles) {
-                            this.generateMedicationHistoryPdf(fileIndex + 1);
-                        } else {
-                            this.isLoading = false;
-                        }
-                    });
-                });
-            });
+    private downloadPdf() {
+        switch (this.reportType) {
+            case "MED":
+                this.medicationHistoryReport.generatePdf();
+                break;
+            case "MSP":
+                this.mspVisitsReport.generatePdf();
+                break;
+        }
     }
 
     private mounted() {
         this.logger = container.get<ILogger>(SERVICE_IDENTIFIER.Logger);
         this.loadName();
-    }
-
-    private fetchMedicationStatements(protectiveWord?: string) {
-        this.isDataLoaded = false;
-        this.isLoading = true;
-
-        this.getMedicationStatements({
-            hdid: this.user.hdid,
-            protectiveWord: protectiveWord,
-        })
-            .then((results) => {
-                if (results.resultStatus == ResultType.Success) {
-                    this.protectiveWordAttempts = 0;
-                    this.medicationStatementHistory = results.resourcePayload;
-                    // Required for the sample page
-                    this.medicationStatementHistoryPage = this.medicationStatementHistory.slice(
-                        0,
-                        50
-                    );
-                    this.sortEntries();
-                    this.isDataLoaded = true;
-                } else if (results.resultStatus == ResultType.Protected) {
-                    this.protectiveWordModal.showModal();
-                    this.protectiveWordAttempts++;
-                } else {
-                    this.logger.error(
-                        "Error returned from the medication statements call: " +
-                            JSON.stringify(results.resultError)
-                    );
-                    this.addError(
-                        ErrorTranslator.toBannerError(
-                            "Fetch Medications Error",
-                            results.resultError
-                        )
-                    );
-                }
-            })
-            .catch((err) => {
-                this.logger.error(err);
-                this.addError(
-                    ErrorTranslator.toBannerError(
-                        "Fetch Medications Error",
-                        err
-                    )
-                );
-            })
-            .finally(() => {
-                this.isLoading = false;
-            });
-    }
-
-    private hasGenerated() {
-        this.logger.debug("finished generating Medication History PDF...");
-        this.isLoading = false;
-    }
-
-    private sortEntries() {
-        this.medicationStatementHistoryPage.sort((a, b) =>
-            a.dispensedDate > b.dispensedDate
-                ? -1
-                : a.dispensedDate < b.dispensedDate
-                ? 1
-                : 0
-        );
-    }
-
-    private onProtectiveWordSubmit(value: string) {
-        this.fetchMedicationStatements(value);
-    }
-
-    private onProtectiveWordCancel() {
-        // Does nothing as it won't be able to fetch pharmanet data.
-        this.logger.debug("protective word cancelled");
     }
 
     private loadName(): void {
@@ -229,19 +67,11 @@ export default class ReportsView extends Vue {
             }
         });
     }
-
-    private onReportTypeChanged() {
-        this.isDataLoaded = false;
-        if (this.reportType == "MED") {
-            this.fetchMedicationStatements();
-        }
-    }
 }
 </script>
 
 <template>
     <div>
-        <LoadingComponent :is-loading="isLoading"></LoadingComponent>
         <div class="my-3 fluid">
             <b-row>
                 <b-col
@@ -258,7 +88,6 @@ export default class ReportsView extends Vue {
                                     v-model="reportType"
                                     data-testid="reportType"
                                     :options="reportTypeOptions"
-                                    @change="onReportTypeChanged"
                                 >
                                 </b-form-select>
                             </b-col>
@@ -267,7 +96,7 @@ export default class ReportsView extends Vue {
                                     variant="primary"
                                     data-testid="exportRecordBtn"
                                     class="mx-auto mt-1 d-block"
-                                    :disabled="!isDataLoaded"
+                                    :disabled="!reportType"
                                     @click="showConfirmationModal"
                                 >
                                     Export Records
@@ -276,7 +105,7 @@ export default class ReportsView extends Vue {
                         </b-row>
                     </div>
                     <div
-                        v-if="reportType == 'MED' && isDataLoaded"
+                        v-if="reportType == 'MED'"
                         data-testid="medicationReportSample"
                     >
                         <b-col>
@@ -284,9 +113,24 @@ export default class ReportsView extends Vue {
                                 <div class="scale">
                                     <div ref="report">
                                         <MedicationHistoryReportComponent
-                                            :medication-statement-history="
-                                                medicationStatementHistoryPage
-                                            "
+                                            ref="medicationHistoryReport"
+                                            :name="fullName"
+                                        />
+                                    </div>
+                                </div>
+                            </div>
+                        </b-col>
+                    </div>
+                    <div
+                        v-else-if="reportType == 'MSP'"
+                        data-testid="mspVisitsReportSample"
+                    >
+                        <b-col>
+                            <div class="mx-auto sample">
+                                <div class="scale">
+                                    <div ref="report">
+                                        <MSPVisitsReportComponent
+                                            ref="mspVisitsReport"
                                             :name="fullName"
                                         />
                                     </div>
@@ -322,14 +166,7 @@ export default class ReportsView extends Vue {
             ref="messageModal"
             title="Sensitive Document Download"
             message="The file that you are downloading contains personal information. If you are on a public computer, please ensure that the file is deleted before you log off."
-            @submit="generateMedicationHistoryPdf"
-        />
-        <ProtectiveWordComponent
-            ref="protectiveWordModal"
-            :error="protectiveWordAttempts > 1"
-            :is-loading="isLoading"
-            @submit="onProtectiveWordSubmit"
-            @cancel="onProtectiveWordCancel"
+            @submit="downloadPdf"
         />
     </div>
 </template>
